refactor: 유지보수성 향상을 위한 패키지 구조 변경
