+++ app/src/main/java/woowacourse/movie/domain/model/DateType.kt
@@ -1,4 +1,4 @@
-package woowacourse.movie
+package woowacourse.movie.domain.model
 
 import java.time.DayOfWeek.SATURDAY
 import java.time.DayOfWeek.SUNDAY
