+++ app/src/main/java/woowacourse/movie/domain/model/MovieTime.kt
@@ -1,4 +1,4 @@
-package woowacourse.movie
+package woowacourse.movie.domain.model
 
 import android.os.Parcelable
 import kotlinx.parcelize.Parcelize
