+++ app/src/main/java/woowacourse/movie/presentation/bookingcomplete/BookingCompleteActivity.kt
@@ -1,4 +1,4 @@
-package woowacourse.movie
+package woowacourse.movie.presentation.bookingcomplete
 
 import android.content.Context
 import android.content.Intent
@@ -11,6 +11,8 @@ import androidx.activity.enableEdgeToEdge
 import androidx.appcompat.app.AppCompatActivity
 import androidx.core.view.ViewCompat
 import androidx.core.view.WindowInsetsCompat
+import woowacourse.movie.R
+import woowacourse.movie.domain.model.BookingInfo
 
 class BookingCompleteActivity : AppCompatActivity() {
     private val bookingInfo: BookingInfo by lazy { getMovieInfoIntent() }
