+++ app/src/main/res/layout/activity_booking_complete.xml
@@ -5,7 +5,7 @@
     android:id="@+id/activity_booking_complete"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".BookingCompleteActivity">
+    tools:context=".presentation.bookingcomplete.BookingCompleteActivity">
 
     <TextView
         android:id="@+id/tv_booking_complete_cancel_notice"
