+++ app/src/main/res/layout/activity_movies.xml
@@ -5,7 +5,7 @@
     android:id="@+id/activity_movies"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".MoviesActivity">
+    tools:context=".presentation.movies.MoviesActivity">
 
     <ListView
         android:id="@+id/lv_movies"
